Move `PostRequestOperation` out of `BackendOperations`
