+++ SDK/Runtime/PostRequestOperation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe76c14687d2b2542b596d4d78ee3459
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
